pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adventure-game-jetson"
version = "0.1.0"
description = "Standalone Jetson-friendly replacement scaffold for the ROS adventure game."
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = {text = "MIT"}
authors = [{name = "Jetson Workspace"}]
dependencies = [
  "numpy>=1.23",
  "opencv-python>=4.8",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "mediapipe>=0.10",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
adventure-game = "adventure_game_jetson.app.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
